requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
crewai==0.30.11
crewai-tools==0.2.6